pyproject.toml
[project]
name = "qna-rag"
version = "0.1.0"
description = "Q and A RAG and document management"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "psycopg2-binary==2.9.10",
    "psycopg-binary==3.2.5",
    "python-dotenv==1.0.1",
    "pgvector==0.3.6",
    "langchain==0.3.19",
    "langchain-ollama==0.2.3",
    "langchain_postgres==0.0.13",
    "langchain-community==0.3.18",
    "pypdf==5.3.0",
    "fastapi==0.115.8",
    "uvicorn==0.34.0",
    "rank_bm25==0.2.2",
    "setuptools==75.8.0"
]